requirements.txt
# Core Pydantic AI dependencies
pydantic-ai>=0.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# OpenAI Provider
openai>=1.0.0

# Supabase Storage
supabase>=2.0.0

# WebSocket and FastAPI
fastapi>=0.104.0
websockets>=12.0
uvicorn>=0.24.0

# Chart Generation
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
seaborn>=0.12.0
pillow>=10.0.0
scipy>=1.10.0

# Async utilities
httpx>=0.25.0
aiofiles>=23.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Monitoring
loguru>=0.7.0